refactor: broken code to different pages for Streamlit
