+++ .gitignore
@@ -1,2 +1,2 @@
 /venv
-/__pycache__/
+__pycache__
