+++ app/pages/1_🔍Initial_Analysis.py
@@ -1,9 +1,7 @@
 import streamlit as st
 import pandas as pd
 
-mpg_data = pd.read_csv(
-    "https://raw.githubusercontent.com/mwaskom/seaborn-data/refs/heads/master/mpg.csv"
-)
+from Data_exploration import mpg_data
 
 st.set_page_config(page_title="Initial Analysis", page_icon="🔍", layout="wide")
 if "expand_code" not in st.session_state:
@@ -23,51 +21,86 @@ with st.sidebar:
         st.session_state.expand_code = False
 
 
-st.markdown("## Reading and cleaning dataset")
+st.markdown("# Extreme values: Max and Min")
+
 st.markdown(
-    """### Initial Observation
-The car dataset has many models of cars produced across different years. The origin and engine performance metrics are listed out.
-"""
-)
-st.markdown(
-    """
-### Check for missing values
-"""
+    "Observe the maximum and minimum extremeties of data and also observe the data to get to know what data we have at hand. For this data, we have different cars, which were produced form 1970 to 1982."
 )
 
-col1, col2 = st.columns(spec=[0.3, 0.7])
+show_code(
+    """mpg_data.loc[mpg_data["horsepower"] == mpg_data["horsepower"].max()]
+mpg_data.loc[mpg_data["horsepower"] == mpg_data["horsepower"].min()]
+mpg_data.loc[mpg_data["acceleration"] == mpg_data["acceleration"].max()]
+mpg_data.loc[mpg_data["acceleration"] == mpg_data["acceleration"].min()]
+mpg_data.loc[mpg_data["mpg"] == mpg_data["mpg"].max()]
+mpg_data.loc[mpg_data["mpg"] == mpg_data["mpg"].min()]
+mpg_data.loc[mpg_data["model_year"] == mpg_data["model_year"].max()]
+mpg_data.loc[mpg_data["model_year"] == mpg_data["model_year"].min()]
+mpg_data.loc[mpg_data["cylinders"] == 3]"""
+)
 
-with col1:
 
-    st.markdown("```mpg_data.isna().sum()```")
-    st.dataframe(mpg_data.isna().sum())
+selected_column = st.selectbox(
+    "Field", options=["horsepower", "mpg", "acceleration", "model_year"]
+)
 
-with col2:
+if selected_column == "model_year":
+    st.write("Newest cars")
+else:
+    st.write("Max value for:", selected_column)
+st.dataframe(mpg_data.loc[mpg_data[selected_column] == mpg_data[selected_column].max()])
 
-    temp = mpg_data[mpg_data.isna().any(axis=1)]
-    selected_column = "horsepower"
-    df_styled = temp.style.set_properties(
-        subset=[selected_column], **{"background-color": "#FF474C"}
+if selected_column == "model_year":
+    st.write("Oldest cars")
+else:
+    st.write("Min value for:", selected_column)
+st.dataframe(mpg_data.loc[mpg_data[selected_column] == mpg_data[selected_column].min()])
+
+
+cylinder_val = st.slider("No of cylinders:", max_value=8, min_value=3, value=4)
+if not cylinder_val == 7:
+    series = (
+        mpg_data.loc[mpg_data["cylinders"] == cylinder_val]
+        .sort_values(by="horsepower", ascending=False)
+        .iloc[0]
     )
-    st.markdown("```mpg_data[mpg_data.isna().any(axis=1)]```")
-    st.dataframe(df_styled)
+    st.dataframe(pd.DataFrame([series]))
+else:
+    st.write("No Cars with 7 cylinders!")
 
+st.markdown("---")
+st.markdown("# Cars in the Dataset")
 
-st.markdown("### Drop/Impute Data")
-st.markdown(
-    "Since there are not many rows that contain null values in horsepower that are not the extremes, I decided to simply drop these data."
-)
-show_code("mpg_data.dropna(inplace=True)")
-
-mpg_data.dropna(inplace=True)
-mpg_data["horsepower"] = pd.to_numeric(mpg_data["horsepower"])
-
-
-st.markdown(
-    """
-## Summary of the Data
-"""
-)
-
-show_code("st.dataframe(mpg_data.describe())")
-st.dataframe(mpg_data.describe())
+col1, col2 = st.columns(spec=[0.5, 0.5])
+with col1:
+    st.image(
+        image="assets/1977_mazda_rx-4.jpg",
+        caption="Powerful 3 cylinder car: 1977 Mazda RX-4",
+    )
+    st.image(
+        image="assets/1980_mazda_rx-7.jpg",
+        caption="Newer gen 3 cylinder car: 1980 Mazda RX-7",
+    )
+    st.image(image="assets/audi_5000.jpg", caption="Powerful 5 cylinder car: Audi 5000")
+    st.image(
+        image="assets/buick_regal_sport_coupe.jpg",
+        caption='Powerful 6 cylinder car: Buick Regal Sport "coupe"',
+    )
+    st.image(image="assets/saab_99le.jpg", caption="Powerful 4 cylinder car: Saab 91LE")
+with col2:
+    st.image(
+        image="assets/harvester_intl_1200D.png",
+        caption="Least fuel efficient in the dataset: Harvester Intl 1200D",
+    )
+    st.image(
+        image="assets/mazda_glc.png",
+        caption="Most fuel efficient in the dataset: Mazda GLC",
+    )
+    st.image(
+        image="assets/pontiac_GP.jpg",
+        caption="Highest Horsepower in the dataset: Pontiac GP",
+    )
+    st.image(
+        image="assets/pugeot_504.png",
+        caption="Highest Acceleration in the dataset: Pugeot 504",
+    )
