+++ app/pages/5_📜Conclusion.py
@@ -1,0 +1,36 @@
+import streamlit as st
+
+st.markdown(
+    """
+# Conclusions
+
+These final graphs give us answers to a few questions we were asking.
+
+## Why would acceleration increase when horsepower is decreasing?
+
+Newer cars have higher acceleration despite lower horsepower. And along with good acceleration, these cars are also fuel efficient.
+
+Turns out the acceleration in cars are increasing beacuse cars are getting lighter, requiring less power to accelerate.
+
+---
+
+## Why did the companies start making lower power, fuel efficient cars?
+
+This data does make it clear that cars do not only need raw horsepower. They can be fast and fuel efficient even with lower horsepower. With fuel emmisions being a major concern and major technological breakthroughs, it is compeletely sensible now that the companies chose to make their cars this way. What is more that the cars could be cheaper with less materials being used and the car would be less expensive to operate for the customers as well all while being better for the environment.
+
+There might be several factors that the data does not capture. The newer cars would have better technology like lighter materials, aerodynamics, better fuel composition and better engineering as well. But the fact is that car companies did make lighter, fuel efficient cars earlier too; only that they chose to keep making newer cars that were lower power and less fuel hungry.
+
+---
+
+## Why are there so many 4 cylinder cars?
+
+This falls in line with the fact that the newer models have on average lower number of cylinders. This also makes us understand as why there are more number of 4 cylinder cars which are the lightest and most fuel efficient, with better stability, lower vibration and possibly the cost.
+
+---
+---
+
+Play around with graphs in the 🛝Playground
+Checkout my Github: (Github-Chocolatewafer)[https://github.com/chocolatewafer]
+
+"""
+)
